requirements.txt
revChatGPT==0.3.1
discord.py==2.1.0